:sparkles: User settings 변경
