+++ src/main/java/bigbrother/slimdealz/SlimDealzApplication.java
@@ -3,8 +3,10 @@ package bigbrother.slimdealz;
 import io.github.cdimascio.dotenv.Dotenv;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class SlimDealzApplication {
 
     public static void main(String[] args) {
