+++ src/main/java/bigbrother/slimdealz/auth/JWTutil.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.auth;
+
+public class JWTutil {
+}
