--- src/main/java/bigbrother/slimdealz/controller/BookmarkController.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.controller;
-
-public class BookmarkController {
-}
