+++ src/main/java/bigbrother/slimdealz/controller/User/BookmarkController.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.controller.User;
+
+public class BookmarkController {
+}
