--- src/main/java/bigbrother/slimdealz/repository/BookmarkRepository.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.repository;
-
-public interface BookmarkRepository {
-}
