--- src/main/java/bigbrother/slimdealz/repository/NotificationRepository.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.repository;
-
-public interface NotificationRepository {
-}
