+++ src/main/java/bigbrother/slimdealz/repository/User/NotificationRepository.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.repository.User;
+
+public interface NotificationRepository {
+}
