+++ src/main/java/bigbrother/slimdealz/repository/User/UserRepository.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.repository.User;
+
+public interface UserRepository {
+}
