--- src/main/java/bigbrother/slimdealz/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.repository;
-
-public interface UserRepository {
-}
