--- src/main/java/bigbrother/slimdealz/service/AuthService.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.service;
-
-public class AuthService {
-}
