--- src/main/java/bigbrother/slimdealz/service/BookmarkService.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.service;
-
-public class BookmarkService {
-}
