--- src/main/java/bigbrother/slimdealz/service/NotificationService.java
@@ -1,4 +1,0 @@
-package bigbrother.slimdealz.service;
-
-public class NotificationService {
-}
