+++ src/main/java/bigbrother/slimdealz/service/User/AuthService.java
@@ -1,0 +1,4 @@
+package bigbrother.slimdealz.service.User;
+
+public class AuthService {
+}
